+++ src/main/java/chess/controller/mapper/TeamMapper.java
@@ -1,7 +1,5 @@
-package chess.view;
+package chess.controller.mapper;
 
-import chess.domain.piece.Empty;
-import chess.domain.piece.PieceType;
 import chess.domain.piece.Team;
 
 import java.util.Arrays;
