+++ src/test/java/chess/view/PieceMapperTest.java
@@ -1,6 +1,7 @@
 package chess.view;
 
 import chess.domain.piece.PieceType;
+import chess.controller.mapper.PieceMapper;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
